refactor(users): move lib/user + lib/userSettings into lib/users/svc
