+++ web/src/sections/AppHealthBanner.tsx
@@ -5,7 +5,7 @@ import useSWR from "swr";
 import { SWR_KEYS } from "@/lib/swr-keys";
 import Modal from "@/refresh-components/Modal";
 import { useCallback, useEffect, useRef, useState } from "react";
-import { logout, refreshToken } from "@/lib/user";
+import { logout, refreshToken } from "@/lib/users/svc";
 import { getSecondsUntilExpiration } from "@opal/time";
 import { NEXT_PUBLIC_CUSTOM_REFRESH_URL } from "@/lib/constants";
 import { Button } from "@opal/components";
